refactor(#8/user): exception 패키지 위치 이동
