+++ src/main/java/org/sopt/post/application/PostService.java
@@ -10,7 +10,7 @@ import org.sopt.post.presentation.response.GetDetailedPostResponse;
 import org.sopt.post.presentation.response.GetSimplePostResponse;
 import org.sopt.post.presentation.response.SearchResultResponse;
 import org.sopt.user.domain.User;
-import org.sopt.user.exception.UserErrorCode;
+import org.sopt.user.presentation.exception.UserErrorCode;
 import org.sopt.user.repository.UserRepository;
 import org.sopt.global.error.BusinessException;
 import org.sopt.post.domain.Post;
